backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
clickhouse-driver==0.2.7
clickhouse-connect==0.7.0
pandas==2.2.1
python-dotenv==1.0.1
httpx==0.27.0
pydantic==2.6.4
pydantic-settings==2.2.1

# Phase 2 — Predictive
prophet==1.1.5
numpy==1.26.4